Make Initial Guess based of 511 keV
I updated the code to make the initial guess looking for 511 keV rather than a pulse at 5110 channels, so it should be more versatile. I also added it so that each sample can have a separate energy spectrum. Right now it does not use energy to line up the channels before calculation, so it isn't done yet, but so far I haven't seen any bugs or crashes.